# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationNotesUpdate": {
        "description": "Schema for updating only the admin_notes field on an application.",
        "properties": {
          "admin_notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes"
          }
        },
        "title": "ApplicationNotesUpdate",
        "type": "object"
      },
      "ApplicationRead": {
        "description": "Schema for reading application data.",
        "properties": {
          "admin_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes"
          },
          "candidate_id": {
            "title": "Candidate Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "growth_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Growth Area"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "pushed_by_admin_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pushed By Admin Id"
          },
          "salary_expectations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Expectations"
          },
          "service_concept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Concept"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "strength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_id",
          "candidate_id",
          "status",
          "admin_notes",
          "service_concept",
          "salary_expectations",
          "strength",
          "growth_area",
          "pushed_by_admin_id",
          "created_at",
          "updated_at"
        ],
        "title": "ApplicationRead",
        "type": "object"
      },
      "ApplicationStatus": {
        "description": "Application (Match) status enumeration.\n\nThe lifecycle (in pipeline order):\n\n    PENDING_ADMIN_REVIEW \u2192 APPROVED_BY_ADMIN \u2192 INTERVIEWING \u2192 OFFER \u2192 HIRED\n                                                            \u2192 REJECTED_BY_COMPANY\n    PENDING_ADMIN_REVIEW / APPROVED_BY_ADMIN \u2192 REJECTED_BY_ADMIN\n                                                 (RS screen-out, pre-employer)\n    any active \u2192 WITHDRAWN (candidate) / JOB_CLOSED (admin closes the job)\n\nRejection is split by actor: ``REJECTED_BY_ADMIN`` is an RS screen-out\nbefore the employer ever sees the candidate; ``REJECTED_BY_COMPANY`` is an\nemployer's decline after review. The semantic groupings below\n(``is_terminal`` / ``is_active`` / ``company_visible`` / ``company_settable``\n/ ``sort_weight``) are the single source of truth \u2014 derive from them rather\nthan re-listing statuses at each call site.",
        "enum": [
          "PENDING_ADMIN_REVIEW",
          "APPROVED_BY_ADMIN",
          "INTERVIEWING",
          "OFFER",
          "HIRED",
          "REJECTED_BY_COMPANY",
          "REJECTED_BY_ADMIN",
          "WITHDRAWN",
          "JOB_CLOSED"
        ],
        "title": "ApplicationStatus",
        "type": "string"
      },
      "ApplicationStatusUpdate": {
        "description": "Schema for admin status updates on an application.\n\nstatus is required \u2014 use ApplicationUpdate for partial updates\nwhere status is optional.",
        "properties": {
          "admin_notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "ApplicationStatusUpdate",
        "type": "object"
      },
      "ApplicationWithDetails": {
        "description": "Schema for application with related job and candidate details.",
        "properties": {
          "admin_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes"
          },
          "ai_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Score"
          },
          "candidate": {
            "$ref": "#/components/schemas/CandidateProfileRead"
          },
          "candidate_id": {
            "title": "Candidate Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "growth_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Growth Area"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "job": {
            "$ref": "#/components/schemas/JobRead"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "pushed_by_admin_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pushed By Admin Id"
          },
          "salary_expectations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Expectations"
          },
          "service_concept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Concept"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "strength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_id",
          "candidate_id",
          "status",
          "admin_notes",
          "service_concept",
          "salary_expectations",
          "strength",
          "growth_area",
          "pushed_by_admin_id",
          "created_at",
          "updated_at",
          "job",
          "candidate"
        ],
        "title": "ApplicationWithDetails",
        "type": "object"
      },
...
      "CandidateProfileRead": {
        "description": "Schema for reading candidate profile data.",
        "properties": {
          "ai_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Score"
          },
          "consent_given_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Given At"
          },
          "consent_policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Policy Version"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_deleted": {
            "description": "Tombstone state, mirroring ``CandidateAdminRead.is_deleted``.\n\nApplications outlive the candidate who submitted them, so every\nconsumer of this schema can be handed a tombstoned row and needs to\ntell one apart without pattern-matching the scrubbed name or the\nsynthetic ``deleted-N@deleted`` address.",
            "readOnly": true,
            "title": "Is Deleted",
            "type": "boolean"
          },
          "is_registered": {
            "readOnly": true,
            "title": "Is Registered",
            "type": "boolean"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "resume_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Path"
          },
          "resume_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Summary"
          },
          "tos_accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tos Accepted At"
          },
          "tos_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tos Version"
          }
        },
        "required": [
          "id",
          "full_name",
          "email",
          "phone",
          "resume_path",
          "linkedin_url",
          "consent_given_at",
          "consent_policy_version",
          "tos_accepted_at",
          "tos_version",
          "created_at",
          "is_registered",
          "is_deleted"
        ],
        "title": "CandidateProfileRead",
        "type": "object"
      },
...
      "CursorPage_ApplicationWithDetails_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationWithDetails"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CursorPage[ApplicationWithDetails]",
        "type": "object"
      },
      "CursorPage_AuditLogRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CursorPage[AuditLogRead]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RS Recruitment API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/applications": {
      "get": {
        "description": "List applications with optional filters, sorted by `sort`/`order`.\n\n`q` case-insensitively substring-matches candidate name/email/phone and\njob title. `sort2`/`order2` add a second sort column as a tiebreaker \u2014 e.g.\n`sort=status&sort2=created_at` groups by status, then by date.\n`include_deleted` (default false) controls whether applications from\ntombstoned candidates appear \u2014 the applications themselves are retained\neither way.\nCursor-paginated.",
        "operationId": "get_applications_api_admin_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "job_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          },
          {
            "in": "query",
            "name": "candidate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_at",
              "enum": [
                "name",
                "created_at",
                "status",
                "score"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "name",
                    "created_at",
                    "status"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort2"
            }
          },
          {
            "in": "query",
            "name": "order2",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order2",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_ApplicationWithDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Applications",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/applications/{application_id}": {
      "delete": {
        "description": "Delete an application record. The candidate profile is preserved.",
        "operationId": "delete_application_endpoint_api_admin_applications__application_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Application Endpoint",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "description": "Get a single application with full details.",
        "operationId": "get_application_detail_api_admin_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationWithDetails"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Application Detail",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/applications/{application_id}/activity": {
      "get": {
        "description": "Activity timeline: audit rows for this application (status changes).",
        "operationId": "get_application_activity_endpoint_api_admin_applications__application_id__activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_AuditLogRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Application Activity Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/applications/{application_id}/notes": {
      "put": {
        "description": "Update only the admin_notes field. Does not change status or send email.",
        "operationId": "update_application_notes_endpoint_api_admin_applications__application_id__notes_put",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationNotesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Application Notes Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/applications/{application_id}/status": {
      "put": {
        "description": "Update application status. Emails (if any) enqueued after commit.",
        "operationId": "update_application_status_endpoint_api_admin_applications__application_id__status_put",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Application Status Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}